feat(SBCount.toString()): implement the feature, "비교 결과 출력하기"
SBCount.toString() create message that will be displayed to the user.
The format of the message is "x볼 x스트라이크" when both the strike count and the ball count are greater than zero. Except when both are zero, a count that is zero will not be displayed.

before printing result of comparing, this process must be performed.